POSIX says timeval's have two long's
SCCS-vsn: sys/sys/time.h 8.5

+++ usr/src/sys/sys/time.h
@@ -4,7 +4,7 @@
  *
  * %sccs.include.redist.c%
  *
- *	@(#)time.h	8.4 (Berkeley) %G%
+ *	@(#)time.h	8.5 (Berkeley) %G%
  */
 
 #ifndef _SYS_TIME_H_
@@ -17,8 +17,8 @@
  * and used in other calls.
  */
 struct timeval {
-	int32_t	tv_sec;		/* seconds */
-	int32_t	tv_usec;	/* and microseconds */
+	long	tv_sec;		/* seconds */
+	long	tv_usec;	/* and microseconds */
 };
 
 /*
